simulator/requirements.txt
Flask==1.1.2
dataclasses-json==0.5.2
psycopg2-binary==2.8.6
python-arango==7.1.0
